requirements.txt
fastapi
uvicorn
sqlalchemy
pymysql
cryptography
python-dotenv
rapidfuzz
openai
pydantic
python-jose[cryptography]
passlib[bcrypt]
python-multipart
streamlit